pyproject.toml
[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"

[project]
name = "pypicoboot"
dynamic = ["version", "readme"]
description = "Pico Boot for Python"
requires-python = ">=3.8"
license = { file = "LICENSE" }

authors = [
  { name = "Pol Henarejos", email = "pol.henarejos@cttc.es" }
]

dependencies = [
    "setuptools",
    "pyusb",
    "esptool",
    "pyserial",
]

classifiers = [
    "Development Status :: 5 - Production/Stable",
    "Environment :: Plugins",
    "Intended Audience :: System Administrators",
    "License :: OSI Approved :: GNU Affero General Public License v3 or later (AGPLv3+)",
    "Programming Language :: Python :: 3.8",
    "Topic :: Security",
    "Topic :: System :: Installation/Setup",
    "Topic :: System :: Networking",
    "Topic :: System :: Systems Administration",
    "Topic :: Utilities",
]

[project.urls]
Homepage = "https://github.com/polhenarejos/pypicoboot"

[tool.setuptools]
packages = [
    "picoboot",
    "picoboot.core",
    "picoboot.tools"
]
include-package-data = true

[tool.setuptools.dynamic]
version = { attr = "picoboot._version.__version__" }
readme = { file = "README.md" }
